+++ summary/summary.go
@@ -1,7 +1,6 @@
 package summary
 
 import (
-	"math"
 	"sort"
 	"strconv"
 	"sync"
@@ -12,67 +11,74 @@ import (
 
 var (
 	AnalysisData  sync.Map
-	ResChanel     = make(chan Res, 50000)
+	ResChanel     = make(chan Res, 20000)
 	RunOverSignal = make(chan int, 1)
 
 	codeDetail  = make(map[int]int)
 	summaryData = SummaryData{
 		CodeDetail:        make(map[string]int),
 		WaitingTimeDetail: make(map[string]int),
-		MinConn:           float64(config.TimeOut),
-		MinDNS:            float64(config.TimeOut),
-		MinDelay:          float64(config.TimeOut),
-		MinReq:            float64(config.TimeOut),
-		MinUseTime:        float64(config.TimeOut),
-		MinRes:            float64(config.TimeOut),
+		MinConn:           int64(config.TimeOut),
+		MinDNS:            int64(config.TimeOut),
+		MinDelay:          int64(config.TimeOut),
+		MinReq:            int64(config.TimeOut),
+		MinUseTime:        int64(config.TimeOut),
+		MinRes:            int64(config.TimeOut),
 	}
 	config    = conf.Conf
-	waitTimes = make([]float64, 0, config.UrlNum)
+	waitTimes = make([]int, 0, config.UrlNum)
+	B         = int64(1000000)
 )
 
 type Res struct {
-	Size         int
-	TimeStamp    int
-	TotalUseTime float64
+	Size         int64
+	TimeStamp    int64
+	TotalUseTime int64
 	Code         int
-	ConnTime     float64
-	DNSTime      float64
-	ReqTime      float64
-	DelayTime    float64
-	ResTime      float64
+	ConnTime     int64
+	DNSTime      int64
+	ReqTime      int64
+	DelayTime    int64
+	ResTime      int64
 }
 
 type SummaryData struct {
 	CompleteRequests      int
 	FailedRequests        int
 	SuccessRequests       int
-	TimeToken             float64
-	TotalDataSize         int
-	AvgDataSize           int
+	TimeTaken             float64
+	TotalDataSize         int64
+	AvgDataSize           int64
 	RequestsPerSec        float64
 	SuccessRequestsPerSec float64
 
-	MinUseTime        float64
-	MaxUseTime        float64
+	MinUseTime        int64
+	MaxUseTime        int64
 	AvgUseTime        float64
+	totolTime         int64
 	CodeDetail        map[string]int
 	WaitingTimeDetail map[string]int
 
-	AvgConn  float64
-	MaxConn  float64
-	MinConn  float64
-	AvgDNS   float64
-	MaxDNS   float64
-	MinDNS   float64
-	AvgReq   float64
-	MaxReq   float64
-	MinReq   float64
-	AvgDelay float64
-	MaxDelay float64
-	MinDelay float64
-	AvgRes   float64
-	MaxRes   float64
-	MinRes   float64
+	AvgConn       float64
+	totalAvgConn  int64
+	MaxConn       int64
+	MinConn       int64
+	AvgDNS        float64
+	totalAvgDNS   int64
+	MaxDNS        int64
+	MinDNS        int64
+	AvgReq        float64
+	totalAvgReq   int64
+	MaxReq        int64
+	MinReq        int64
+	AvgDelay      float64
+	totalAvgDelay int64
+	MaxDelay      int64
+	MinDelay      int64
+	AvgRes        float64
+	totalAvgRes   int64
+	MaxRes        int64
+	MinRes        int64
 }
 
 func HandleRes() {
@@ -86,6 +92,7 @@ func HandleRes() {
 		if summaryData.CompleteRequests == config.UrlNum {
 			close(ResChanel)
 		}
+
 		code := res.Code
 		if _, ok := codeDetail[code]; ok {
 			codeDetail[code]++
@@ -100,53 +107,66 @@ func HandleRes() {
 		} else {
 			summaryData.SuccessRequests++
 		}
-		summaryData.AvgUseTime += res.TotalUseTime
-		summaryData.AvgConn += res.ConnTime
-		summaryData.AvgDNS += res.DNSTime
-		summaryData.AvgDelay += res.DelayTime
-		summaryData.AvgReq += res.ReqTime
-		summaryData.AvgRes += res.ResTime
+		summaryData.totolTime += res.TotalUseTime
+		summaryData.totalAvgConn += res.ConnTime
+		summaryData.totalAvgDNS += res.DNSTime
+		summaryData.totalAvgDelay += res.DelayTime
+		summaryData.totalAvgReq += res.ReqTime
+		summaryData.totalAvgRes += res.ResTime
 
-		summaryData.MinUseTime = math.Min(res.TotalUseTime, summaryData.MinUseTime)
-		summaryData.MinConn = math.Min(res.ConnTime, summaryData.MinConn)
-		summaryData.MinDNS = math.Min(res.DNSTime, summaryData.MinDNS)
-		summaryData.MinDelay = math.Min(res.DelayTime, summaryData.MinDelay)
-		summaryData.MinReq = math.Min(res.ReqTime, summaryData.MinReq)
-		summaryData.MinRes = math.Min(res.ResTime, summaryData.MinRes)
+		summaryData.MinUseTime = tools.MinInt64(res.TotalUseTime, summaryData.MinUseTime)
+		summaryData.MinConn = tools.MinInt64(res.ConnTime, summaryData.MinConn)
+		summaryData.MinDNS = tools.MinInt64(res.DNSTime, summaryData.MinDNS)
+		summaryData.MinDelay = tools.MinInt64(res.DelayTime, summaryData.MinDelay)
+		summaryData.MinReq = tools.MinInt64(res.ReqTime, summaryData.MinReq)
+		summaryData.MinRes = tools.MinInt64(res.ResTime, summaryData.MinRes)
 
-		summaryData.MaxUseTime = math.Max(res.TotalUseTime, summaryData.MaxUseTime)
-		summaryData.MaxConn = math.Max(res.ConnTime, summaryData.MaxConn)
-		summaryData.MaxDNS = math.Max(res.DNSTime, summaryData.MaxDNS)
-		summaryData.MaxDelay = math.Max(res.DelayTime, summaryData.MaxDelay)
-		summaryData.MaxReq = math.Max(res.ReqTime, summaryData.MaxReq)
-		summaryData.MaxRes = math.Max(res.ResTime, summaryData.MaxRes)
-		waitTimes = append(waitTimes, res.TotalUseTime)
-
+		summaryData.MaxUseTime = tools.MaxInt64(res.TotalUseTime, summaryData.MaxUseTime)
+		summaryData.MaxConn = tools.MaxInt64(res.ConnTime, summaryData.MaxConn)
+		summaryData.MaxDNS = tools.MaxInt64(res.DNSTime, summaryData.MaxDNS)
+		summaryData.MaxDelay = tools.MaxInt64(res.DelayTime, summaryData.MaxDelay)
+		summaryData.MaxReq = tools.MaxInt64(res.ReqTime, summaryData.MaxReq)
+		summaryData.MaxRes = tools.MaxInt64(res.ResTime, summaryData.MaxRes)
+		waitTimes = append(waitTimes, int(res.TotalUseTime/B))
 	}
+	urlNum := int64(config.UrlNum)
+	summaryData.AvgDataSize = summaryData.TotalDataSize / urlNum
 
-	summaryData.AvgUseTime = tools.Decimal2(summaryData.AvgUseTime / float64(config.UrlNum))
-	summaryData.AvgConn = tools.Decimal2(summaryData.AvgConn / float64(config.UrlNum))
-	summaryData.AvgDNS = tools.Decimal2(summaryData.AvgDNS / float64(config.UrlNum))
-	summaryData.AvgDelay = tools.Decimal2(summaryData.AvgDelay / float64(config.UrlNum))
-	summaryData.AvgReq = tools.Decimal2(summaryData.AvgReq / float64(config.UrlNum))
-	summaryData.AvgRes = tools.Decimal2(summaryData.AvgRes / float64(config.UrlNum))
-	summaryData.AvgDataSize = summaryData.TotalDataSize / config.UrlNum
+	summaryData.AvgUseTime = tools.Decimal2(float64(summaryData.totolTime / urlNum / B))
+	summaryData.AvgConn = tools.Decimal2(float64(summaryData.totalAvgConn / urlNum / B))
+	summaryData.AvgDNS = tools.Decimal2(float64(summaryData.totalAvgDNS / urlNum / B))
+	summaryData.AvgDelay = tools.Decimal2(float64(summaryData.totalAvgDelay / urlNum / B))
+	summaryData.AvgReq = tools.Decimal2(float64(summaryData.totalAvgReq / urlNum / B))
+	summaryData.AvgRes = tools.Decimal2(float64(summaryData.totalAvgRes / urlNum / B))
+
+	summaryData.MaxConn = summaryData.MaxConn / B
+	summaryData.MaxDNS = summaryData.MaxDNS / B
+	summaryData.MaxDelay = summaryData.MaxDelay / B
+	summaryData.MaxReq = summaryData.MaxReq / B
+	summaryData.MaxRes = summaryData.MaxRes / B
+	summaryData.MaxUseTime = summaryData.MaxUseTime / B
+	summaryData.MinConn = summaryData.MinConn / B
+	summaryData.MinDNS = summaryData.MinDNS / B
+	summaryData.MinDelay = summaryData.MinDelay / B
+	summaryData.MinReq = summaryData.MinReq / B
+	summaryData.MinRes = summaryData.MinRes / B
+	summaryData.MinUseTime = summaryData.MinUseTime / B
 
 	for k, v := range codeDetail {
 		summaryData.CodeDetail[strconv.Itoa(k)] = v
 	}
 
 	t := (float64(config.EndTime-config.StartTime) / 10e8)
-	summaryData.TimeToken = t
+	summaryData.TimeTaken = t
 	summaryData.RequestsPerSec = float64(config.UrlNum) / t
 	summaryData.SuccessRequestsPerSec = float64(summaryData.SuccessRequests) / t
-	sort.Float64s(waitTimes)
-	waitTimesL := float64(len(waitTimes))
+	sort.Ints(waitTimes)
+	waitTimesL := len(waitTimes)
 	tps := []float64{0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99, 0.999, 0.9999}
 	tpsL := len(tps)
 	for i := 0; i < tpsL; i++ {
 		summaryData.WaitingTimeDetail[tools.FloatToPercent(
-			tps[i])] = int(waitTimes[int(waitTimesL*tps[i]-1)])
+			tps[i])] = int(waitTimes[int(float64(waitTimesL)*tps[i]-1)])
 	}
 	Print(summaryData)
 }